Improved memory usage in certain cases
The old behavior of checking most recently added WFMO waits first
in the SetEvent function could lead to excessive memory usage in
cases where SetEvent is never called except when there is already
a wait in progress. This left old, completed waits in the WFMO
RegisteredWaits vector, which could accumulate and take undesired
amounts of memory if no spurious SetEvents came through. So not a
memory leak per se, but undesirable memory greediness all the same.

This commit swaps the vector of WFMO waits for a deque, and signals
the WFMO waits in a first come, first serve method; this behavior
also has the advantage of being more in line with the WIN32 event
implementation.

Hat-tip to Ashwin Chandra <achandra@panologic.com> for pointing out
the excessive memory usage leading us to research this further and
change the behavior to the aforementioned first come, first serve.