More consistent error handling and assertion
The problem with code dealing with mutexes, conditional variables,
and other locking implements is that if the lock operations fail
at certain points, the code can be left in an inconsistent state,
and there is not necessarily a way of notifying the calling code
that this has happened if we wish to stick to the pthread coding
conventions.

This commit introduces assert(false) for all operations that are
unrecoverable errors - i.e. resulting in an inconsistent state
from which any calling code (users of the pevents library) cannot
recover. Failures that occur in the platform's underlying locking
functions that take place at a point where the pevent state is
still consistent and known simply clean up and pass back the error
to the calling code.

This commit adds error checking for *all* the underlying locking
calls.