Added option to disable error checking on underlying pthread funcs
Error checking on the underlying pthread functions has been grouped
into an #ifdef. Compiling with PTHREADCHK will enable checking the
return values of all usages of the platform's pthread functions.
Compiling without PTHREADCHK will improve performance by removing
all branches related to error-checking results of pthread
functions.

If you are certain your code will not be operating close to the OS'
limits in terms of memory or resources, it should be OK to disable
PTHREADCHK in exchange for a slight boost in performance.

PTHREADCHK should always be enabled for debug builds.