Implemented early-cleanup for expired waits
When a new WaitForMultipleEvents call is made, any expired waits
still queued/listed in the RegisteredWaits deque for the events
passed to the WFMO function will be cleaned up.

Previously, in cases where there is an event that is always waited
on but almost never triggered in a WFMO wait with waitAll set to
false and more than one event passed in the events array, the
resources used for the pending waits would accumulate. Depending
on just how often WFMO is called before this event is triggered,
the total memory usage could be rather high.

While this incurs a slight performance penalty (looping over
other pending waits, acquiring locks for other waits for these
events), this isn't an issue that can be neglected. I believe the
mutex locking in the early cleanup remove_if predicate might be
replacable with some nifty CAS functions, and will look into that
later.