Added implementation of PulseEvent(), disabled by default
Even though PulseEvent() is and always has been completely broken
and unsafe to rely on for deadlock-free code [0], I've decided to
include it in pevents as an optional feature via -DPULSE as some
libraries being ported from Windows to *nix are using PulseEvent()
and will want to implement similar behavior in pevents instead of
rewriting the code to actually use sound synchronization methods.

I figure writing as close to a correct PulseEvent() implementation
that behaves "as right/wrong" as the Win32 implementation to be a
better scenario than a fully broken PulseEvent() being hacked on
to pevents.

[0] Excellent discussion on PulseEvent() and why it's broken:
http://blogs.msdn.com/b/oldnewthing/archive/2005/01/05/346888.aspx