Cross-platform support, now compiles on Windows too
With this commit, pevents will compile on Windows and posix. On
Windows, it will use the native CreateEvent() functions, as
defined in Win32.cpp. On other platforms, it will use the previous
pthreads abstraction layer in pevents.cpp

Because Windows.h (Winbase.h or Synchapi.h) define CreateEvent to
either CreateEventA or CreateEventW, our function will be renamed
too, but that should be fine.

The error codes for Create/Set/Reset/Wait are platform-dependant.
WAIT_TIMEOUT can be used to check for a timeout on all platforms,
on *nix it is defined to be ETIMEDOUT.

WaitForEvent/WaitForMultipleEvents support 64-bit timeoouts, the
original, native Win32 functions only support 32-bit timeouts. Yay!