Initial conversion attempt to pure C code
The functionality is now complete without relying on the standard
library, but there is a race condition that is caused by the change
to the event triggering mechanism.

Currently, a SetEvent call co-inciding with a WFMO call terminating
can cause a race condition where a WFMO in the event's registered
waits is de-initialized just after the event attempts to trigger it
when enumerating its registered events.

Either need to go back to original mechanism of registering/de-
registering waits or remove wait elision from the SetEvent call
entirely.