Restructured code to use reference counting and changed cleanup
Instead of trying to clean everything up in the calling thread,
now using ref counting to let events clean themselves up on
SetEvent calls (like in the original C++ version of the pevents
code).

While the idea of eliminating ref counting was alluring, it did
not make sense and introduced too many complications due to the
need to track which pointers to WFMO instances were valid or not
in the RegisteredEvents list/queue.

After further reflection, this is likely to result in better
real-world performance, as it is the assumption of this developer
that it is more important for calls to WFSO/WFMO to return faster
than it is for SetEvent and  co. to terminate earlier. Either way,
the difference should not be too bad and the waits will never
block on user activity - the only mutexes waited on are private to
the pevents library itself.