Calling pthread_mutex_unlock before WFMO->Destroy()
As discussed/discovered in #2, the behavior of pthread_mutex_destroy
is not defined if the mutex is held when it is destroyed. We need
to release the mutex first.

This simplifies the code flow somewhat; it's not our job to protect
the user from calling SetEvent/WaitEvent at the same time as a
call to DestroyEvent.