Rename neosmart_event_t to pevent_t
It also doesn't seem correct to use `pthread_event_t` as the name if the
event object is not part of the `pthread` library proper, even if it is
built atop of it. Furthermore, there's no need for the name to be so
long and wordy if we have the advantage of C++ namespacing, since
there's no risk of a name collision.