Drop zero-timeout pthread_mutex_trylock() optimization
Previously, if the mutex protecting the inner state of the event object
were locked by another thread, the early abort in `WaitForEvent()` with
a zero-ms timeout would have translated that to a `WAIT_TIMEOUT` result,
even if the application logic is such that the event in question is
always set (e.g. simultaneous calls to `WaitForEvent()` and `SetEvent()`
on an already-set event).

This patch makes `event->State` an atomic that can be used as a
synchronization point (currently in addition to rather than instead of
`event->Mutex`).

Note: Unlocking the spinlock (i.e. resetting the event) is done with
release semantics rather than using `std::memory_order_relaxed` due to
ambiguities in the C++ spec that make it unclear whether or not
consecutive Set/Reset blocks could be interleaved with the latter; see
the discussion preceding [0] for reference.

Closes #18.

[0]: https://old.reddit.com/r/cpp/comments/g84bzv/c/fpua2yq/